test(codo): DOMA-13250 CI problems fixes again
